+++ src/CleanRegex/Exception/EmptyOptionalException.php
@@ -1,0 +1,6 @@
+<?php
+namespace TRegx\CleanRegex\Exception;
+
+class EmptyOptionalException extends \Exception
+{
+}