+++ test/Feature/CleanRegex/Match/group/asInt/_unmatched/MatchGroupIntStreamTest.php
@@ -2,44 +2,33 @@
 namespace Test\Feature\CleanRegex\Match\group\asInt\_unmatched;
 
 use PHPUnit\Framework\TestCase;
+use Test\Utils\AssertsOptionalEmpty;
 use Test\Utils\ExactExceptionMessage;
 use Test\Utils\Functions;
-use TRegx\CleanRegex\Exception\NoSuchStreamElementException;
-use TRegx\CleanRegex\Exception\SubjectNotMatchedException;
 
 class MatchGroupIntStreamTest extends TestCase
 {
-    use ExactExceptionMessage;
+    use ExactExceptionMessage, AssertsOptionalEmpty;
 
     /**
      * @test
      */
-    public function shouldThrow_asInt()
+    public function shouldReturnEmptyOptional()
     {
-        // given
-        $optional = pattern('(Foo)')->match('Bar')->group(1)->asInt()->findFirst(Functions::fail());
-
-        // then
-        $this->expectException(SubjectNotMatchedException::class);
-        $this->expectExceptionMessage('Expected to get group #1 as integer from the first match, but subject was not matched at all');
-
         // when
-        $optional->get();
+        $optional = pattern('(Foo)')->match('Bar')->group(1)->asInt()->findFirst(Functions::fail());
+        // then
+        $this->assertOptionalEmpty($optional);
     }
 
     /**
      * @test
      */
-    public function shouldThrow_asInt_distinct()
+    public function shouldReturnEmptyOptional_distinct()
     {
-        // given
-        $optional = pattern('(Foo)')->match('Bar')->group(1)->asInt()->distinct()->findFirst(Functions::fail());
-
-        // then
-        $this->expectException(NoSuchStreamElementException::class);
-        $this->expectExceptionMessage('Expected to get group #1 as integer from the first match, but subject was not matched at all');
-
         // when
-        $optional->get();
+        $optional = pattern('(Foo)')->match('Bar')->group(1)->asInt()->distinct()->findFirst(Functions::fail());
+        // then
+        $this->assertOptionalEmpty($optional);
     }
 }