+++ test/Feature/CleanRegex/Match/stream/findNth/MatchPatternTest.php
@@ -2,10 +2,8 @@
 namespace Test\Feature\CleanRegex\Match\stream\findNth;
 
 use PHPUnit\Framework\TestCase;
-use Test\Utils\ExactExceptionMessage;
-use Test\Utils\ExampleException;
+use Test\Utils\AssertsOptional;
 use Test\Utils\Functions;
-use TRegx\CleanRegex\Exception\NoSuchNthElementException;
 
 /**
  * @covers \TRegx\CleanRegex\Match\MatchPattern
@@ -13,7 +11,7 @@ use TRegx\CleanRegex\Exception\NoSuchNthElementException;
  */
 class MatchPatternTest extends TestCase
 {
-    use ExactExceptionMessage;
+    use AssertsOptional;
 
     /**
      * @test
@@ -21,12 +19,11 @@ class MatchPatternTest extends TestCase
     public function shouldThrow_findNth_forUnmatchedSubject()
     {
         // given
-        $optional = pattern('Foo')->match('Bar')->stream()->findNth(5);
-        // then
-        $this->expectException(NoSuchNthElementException::class);
-        $this->expectExceptionMessage('Expected to get the 5-nth stream element, but the subject backing the stream was not matched');
+        $stream = pattern('Foo')->match('Bar')->stream();
         // when
-        $optional->get();
+        $optional = $stream->findNth(5);
+        // then
+        $this->assertOptionalEmpty($optional);
     }
 
     /**
@@ -35,12 +32,11 @@ class MatchPatternTest extends TestCase
     public function shouldThrow_findNth_forStreamUnderflow()
     {
         // given
-        $optional = pattern('\d+')->match('12 13 14')->stream()->findNth(5);
-        // then
-        $this->expectException(NoSuchNthElementException::class);
-        $this->expectExceptionMessage('Expected to get the 5-nth stream element, but the stream has 3 element(s)');
+        $stream = pattern('\d+')->match('12 13 14')->stream();
         // when
-        $optional->get();
+        $optional = $stream->findNth(5);
+        // then
+        $this->assertOptionalEmpty($optional);
     }
 
     /**
@@ -49,40 +45,10 @@ class MatchPatternTest extends TestCase
     public function shouldThrow_findNth_forEmptyStream()
     {
         // given
-        $optional = pattern('Foo')->match('Foo')->stream()->filter(Functions::constant(false))->findNth(5);
+        $stream = pattern('Foo')->match('Foo')->stream()->filter(Functions::constant(false));
+        // when
+        $optional = $stream->findNth(5);
         // then
-        $this->expectException(NoSuchNthElementException::class);
-        $this->expectExceptionMessage('Expected to get the 5-nth stream element, but the stream has 0 element(s)');
-        // when
-        $optional->get();
-    }
-
-    /**
-     * @test
-     */
-    public function shouldThrow_findNth_orThrow_WithCustomException()
-    {
-        // then
-        $this->expectException(ExampleException::class);
-        // when
-        pattern('\d+')->match('12 13 14')->stream()->findNth(5)->orThrow(new ExampleException());
-    }
-
-    /**
-     * @test
-     */
-    public function shouldCall_findNth_forUnmatchedSubject()
-    {
-        // when
-        pattern('(?<pepsi>Foo)')->match('Bar')->stream()->findNth(0)->orElse(Functions::pass());
-    }
-
-    /**
-     * @test
-     */
-    public function shouldCall_findNth_forStreamUnderflow()
-    {
-        // given
-        pattern('(?<pepsi>\d+)')->match('Foo 14')->stream()->findNth(1)->orElse(Functions::pass());
+        $this->assertOptionalEmpty($optional);
     }
 }