+++ test/Utils/AssertsOptional.php
@@ -1,0 +1,7 @@
+<?php
+namespace Test\Utils;
+
+trait AssertsOptional
+{
+    use AssertsOptionalEmpty, AssertsOptionalPresent;
+}